pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcript-python-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
  "python-telegram-bot>=21.6",
  "Flask>=3.0.0",
  "youtube-transcript-api>=0.6.2",
  "notion-client>=2.2.1",
  "PyYAML>=6.0.2",
  "langdetect>=1.0.9",
]

[project.scripts]
transcript-bot-polling = "transcript_python_bot.main:main"
transcript-cli = "transcript_python_bot.cli:main"
transcript-set-webhook = "transcript_python_bot.set_webhook:main"
transcript-delete-webhook = "transcript_python_bot.delete_webhook:main"

[tool.setuptools.packages.find]
where = ["src"]

